rename module, in line with other renamings
